regulatory-compliance-agent/requirements-agent.txt
bedrock-agentcore-starter-toolkit>=0.1.0
strands-agents>=0.1.0
requests>=2.31.0
uvicorn>=0.24.0
pydantic>=2.5.0
